Support for resonant platform (not ready): 1) ioboard supports elongating ttl pulse 2) setup configuration updated 3) additional python 3 compatibility 4) EOL problem eliminated

+++ engine/hardware_interface/mesc_interface.py
@@ -1,32 +1,66 @@
-import socket,subprocess,psutil,time,json,unittest,os
+import socket,subprocess,psutil,time,json,unittest,os,sys
 from visexpman.engine.generic import fileop
+if 1:
+    try:
+        import mescapi
+        import PyQt5.QtCore as QtCore
+        use_proxy=False
+    except ImportError:
+        use_proxy=True
+else:
+    use_proxy=True
 
 class MescapiInterface(object):
-    def __init__(self, port=11000, command_buffer_size=1024*512, timeout=1.0,debug=False):
-        self.command_buffer_size=command_buffer_size
-        apiserverpath=os.path.join(fileop.visexpman_package_path(), 'engine', 'external', 'mesc', 'MEScApiServer.exe')
-        self.serverp=subprocess.Popen([apiserverpath, str(port)],shell=not debug)
-        self.serverpp=psutil.Process(self.serverp.pid)
-        self.clientsocket=socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        #self.clientsocket.settimeout(timeout)
-        try:
-            self.clientsocket.connect(('localhost', port))
-        except:
-            self.serverpp.kill()
-            raise IOError('Cannot connect to MEScApiServer')
+    def __init__(self, mesc_address='localhost', port=11000, command_buffer_size=1024*512, timeout=1.0,debug=False):
+        if use_proxy:
+            self.command_buffer_size=command_buffer_size
+            apiserverpath=os.path.join(fileop.visexpman_package_path(), 'engine', 'external', 'mesc', 'MEScApiServer.exe')
+            self.serverp=subprocess.Popen([apiserverpath, str(port)],shell=not debug)
+            self.serverpp=psutil.Process(self.serverp.pid)
+            self.clientsocket=socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            #self.clientsocket.settimeout(timeout)
+            try:
+                self.clientsocket.connect((mesc_address, port))
+            except:
+                self.serverpp.kill()
+                raise IOError('Cannot connect to MEScApiServer')
+        else:
+            self.manager = mescapi.APIClientManager()
+            done=self.manager.webSocketConnect(QtCore.QUrl('ws://{0}:8888'.format(mesc_address)))
+            if done == False:
+                raise IOError('Cannot connect to MEScApiServer')
+            self.client=self.manager.getClientListModel().getClient(0)
+            loginParser=self.client.login('default','defaultpw')
+            resultCode=loginParser.getResultCode()
+            if resultCode > 0:
+                raise RuntimeError (loginParser.getErrorText())
 
     def request(self,cmd):
-        try:
-            self.clientsocket.send(cmd)
-        except:
-            self.serverpp.kill()
-            raise IOError('Command cannot be sent to MEScApiServer')
-        try:
-            data = self.clientsocket.recv(self.command_buffer_size)
-        except:
-            self.serverpp.kill()
-            raise IOError('MEScApiServer does not respond')
+        if use_proxy:
+            try:
+                if sys.version_info.major==3:
+                    cmd=bytearray(cmd,'utf-8')
+                self.clientsocket.send(cmd)
+            except:
+                self.serverpp.kill()
+                raise IOError('Command cannot be sent to MEScApiServer')
+            try:
+                data = self.clientsocket.recv(self.command_buffer_size)
+            except:
+                self.serverpp.kill()
+                raise IOError('MEScApiServer does not respond')
+        else:
+            parser=self.client.sendJSCommand(cmd)
+            res=parser.getResultCode()
+            if res > 0:
+                raise RuntimeError(parser.getErrorText())
+            else:	
+                data=parser.getJSEngineResult()
         self.data=data
+        if isinstance(data, bool):
+            return data
+        elif not isinstance(data, str):
+            data=data.decode('utf-8')
         try:
             return json.loads(data.replace('\xb5','u'))#unicode is used for um
         except ValueError:
@@ -54,7 +88,7 @@ class MescapiInterface(object):
         '''
         get microscope state
         '''
-        return self.request('MEScMicroscope.getAcquisitionState()').values()[0]
+        return self.request('MEScMicroscope.getAcquisitionState()')['microscopeState']
     
     def save(self, filename):
         '''
@@ -65,31 +99,34 @@ class MescapiInterface(object):
         self.serverpp.kill()
     
     def close(self):
-        self.clientsocket.close()
-        try:
-            self.serverpp.kill()
-        except psutil.NoSuchProcess:
-            pass
+        if use_proxy:
+            self.clientsocket.close()
+            try:
+                self.serverpp.kill()
+            except psutil.NoSuchProcess:
+                pass
+        else:
+            self.manager.closeConnection(self.client)
         
 class TestMesc(unittest.TestCase):
     def test_01_simple_use(self):
         from visexpman.engine.generic import introspect
         m=MescapiInterface(debug=True)
         with introspect.Timer():
-            m.request('MEScFile.getMescState();')
+            m.request('MEScMicroscope.getAcquisitionState();')
         m.close()
         
     def test_02_multiple_commands(self):
         m=MescapiInterface(debug=True)
         for i in range(10):
-            self.assertEqual(m.request('MEScMicroscope.getAcquisitionState();').values(),['Ready'])
+            self.assertEqual([i for i in m.request('MEScMicroscope.getAcquisitionState();').values()],['Ready'])
             self.assertEqual(m.request('MEScFile.echo(\'TEST\');'), 'MEScFile:TEST')
         m.close()
         
     def test_03_reinstantiate(self):
         for i in range(10):
             m=MescapiInterface(debug=True)
-            self.assertEqual(m.request('MEScMicroscope.getAcquisitionState();').values(),['Ready'])
+            self.assertEqual(m.get_state(),'Ready')
             m.close()
             m=MescapiInterface()
             self.assertEqual(m.request('MEScFile.echo(\'TEST\');'), 'MEScFile:TEST')
@@ -97,9 +134,10 @@ class TestMesc(unittest.TestCase):
         
     def test_04_invalid_command(self):
         m=MescapiInterface(debug=True)
-        self.assertEqual(m.request('invalidcmd();')['MESc Error code'],4)
+        self.assertRaises(RuntimeError, m.request, 'invalidcmd();')
         m.close()
         
+    @unittest.skip('')
     def test_05_high_load(self):
         '''
         Multiple commands with lot of data
@@ -126,10 +164,12 @@ class TestMesc(unittest.TestCase):
     def test_06_measurement_start(self):
         m=MescapiInterface(debug=True)
         self.assertRaises(IOError, m.stop)
-        self.assertTrue(m.start()['succeeded'])
-        self.assertTrue(m.stop()['succeeded'])
+        self.assertTrue(m.start())
+        time.sleep(3)
+        self.assertTrue(m.stop())
         m.close()
         
         
 if __name__=='__main__':
+    app = QtCore.QCoreApplication(sys.argv)
     unittest.main()
